@@ -8,36 +8,36 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.1.4",
-    "@angular/common": "^7.1.4",
-    "@angular/compiler": "^7.1.4",
-    "@angular/core": "^7.1.4",
+    "@angular/animations": "^7.2.1",
+    "@angular/common": "^7.2.1",
+    "@angular/compiler": "^7.2.1",
+    "@angular/core": "^7.2.1",
     "@angular/fire": "^5.1.1",
-    "@angular/forms": "^7.1.4",
-    "@angular/http": "^7.1.4",
-    "@angular/material": "^7.2.0",
-    "@angular/platform-browser": "^7.1.4",
-    "@angular/platform-browser-dynamic": "^7.1.4",
-    "@angular/pwa": "^0.11.4",
-    "@angular/router": "^7.1.4",
-    "@angular/service-worker": "^7.1.4",
+    "@angular/forms": "^7.2.1",
+    "@angular/http": "^7.2.1",
+    "@angular/material": "^7.2.1",
+    "@angular/platform-browser": "^7.2.1",
+    "@angular/platform-browser-dynamic": "^7.2.1",
+    "@angular/pwa": "^0.12.2",
+    "@angular/router": "^7.2.1",
+    "@angular/service-worker": "^7.2.1",
     "@ngxs/storage-plugin": "^3.3.4",
     "@ngxs/store": "^3.3.4",
-    "core-js": "^2.6.1",
-    "firebase": "^5.7.1",
+    "core-js": "^2.6.2",
+    "firebase": "^5.8.0",
     "hammerjs": "^2.0.8",
     "rxjs": "^6.3.3",
-    "zone.js": "^0.8.26"
+    "zone.js": "^0.8.28"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.11.4",
-    "@angular/cdk": "^7.2.0",
-    "@angular/cli": "^7.1.4",
-    "@angular/compiler-cli": "^7.1.4",
-    "@angular/language-service": "^7.1.4",
+    "@angular-devkit/build-angular": "^0.12.2",
+    "@angular/cdk": "^7.2.1",
+    "@angular/cli": "^7.2.2",
+    "@angular/compiler-cli": "^7.2.1",
+    "@angular/language-service": "^7.2.1",
     "@types/node": "^10.12.18",
     "codelyzer": "^4.5.0",
-    "tslint": "^5.12.0",
-    "typescript": "^3.1.6"
+    "tslint": "^5.12.1",
+    "typescript": "^3.2.4"
   }
 }
