@@ -18,7 +18,7 @@ export const environment = {
     config: {
       name: 'whiteFire',
       automaticDataCollectionEnabled: true,
-      timestampsInSnapshots: true,
+      // timestampsInSnapshots: true,
     }
   }
 };
